Update CC:T integration to use 1.20.4 features
 - Switch to NF-style capabilities.
 - Remove TaskCallback, as that is now exposed in the public API.
 - Fix arguments not being marked as escaping when running off-thread.
 - Replace ItemStack/FluidStack conversion with CC's detail providers.
