+++ gradle.properties
@@ -9,6 +9,6 @@ version_neoforge=20.4.196
 version_neogradle=7.0.96
 version_ie=11.2.0
 version_jei=17.0.0.30
-version_cctweaked=1.102.0-SNAPSHOT
+version_cctweaked=1.110.2
 version_buildnumber=173
 modid=immersiveengineering
