+++ src/main/java/blusunrize/immersiveengineering/common/util/compat/computers/cctweaked/GenericPeripheral.java
@@ -63,6 +63,13 @@ public class GenericPeripheral<T extends BlockEntity> implements IDynamicPeriphe
 		return creator.getName();
 	}
 
+	@Nullable
+	@Override
+	public Object getTarget()
+	{
+		return object;
+	}
+
 	@Override
 	public boolean equals(@Nullable IPeripheral other)
 	{
